Implementing tracer on users

+++ handler/security/payload/request_edit_password.go
@@ -1,6 +1,6 @@
 package payload
 
 type RequestEditPassword struct {
-	UserID   int    `json:"user_id" validate:"required"`
+	UserID   int64  `json:"user_id" validate:"required"`
 	Password string `json:"password" validate:"required,min=6,max=12"`
 }
